chore: prepare to bump to `0.7.2` ver (#201)
* refactor: some refactors

* fix(ffmpeg-engine): invalid time unit

* fix(kotlin-engine): throw FlvDataErrorException instead

* refactor(download): update streamer status if db write succeed

* refactor(flv-processing): migrate to use kotlinx-io

* refactor(flv-processing): migrate amf parsing to kotlinx-io and drop amf3 support

* fix(download): streamer live end status not updated

* refactor(flv-processing): migrate `AVCSequenceHeaderParser` to use kotlinx-io

* refactor(download): do not update status if equals

* fix(huya-test): remove missing field

* fix: add missing callback

* refactor(ffmpeg-engine): use `FFmpegDownloadEngine` logger

* refactor(douyin-signature): use `DouyinExtractor` logger

* refactor(download): some refactors

* refactor(flv-processing): throw `FlvHeaderErrorException` if header is incomplete

* fix(ffmpeg-engine): skip 0 duration segment time parameter

* fix(download): delete danmu file if download file do not exist

* refactor(download): use streamer context instead

* refactor(string-utils): delete useless `replaceTimestamps` flag

* refactor(extractor): commonize extractor cookies settings

* refactor(download): commonize logging context

* fix(douyin): invalid filter order

* fix(download): build path string from generic path

* fix(kotlin-engine): remove `PART` prefix on edge case

* refactor(download)!: disable segmentation if  `0`  set

+++ flv-processing/src/main/kotlin/github/hua0512/flv/operators/FlvConcat.kt
@@ -69,7 +69,7 @@ private const val MAX_DURATION = 20_000
  */
 internal fun Flow<FlvData>.concat(context: StreamerContext): Flow<FlvData> = flow {
 
-  var delta = 0L
+  var delta = 0
 
   var action: FlvConcatAction = NOOP
 
@@ -155,15 +155,15 @@ internal fun Flow<FlvData>.concat(context: StreamerContext): Flow<FlvData> = flo
   }
 
   fun updateDeltaDuplicated(tag: FlvTag) {
-    delta = (lastTags.last() as FlvTag).header.timestamp - tag.header.timestamp
+    delta = ((lastTags.last() as FlvTag).header.timestamp - tag.header.timestamp).toInt()
   }
 
   fun updateDeltaNonDuplicated(tag: FlvTag) {
-    delta = (lastTags.last() as FlvTag).header.timestamp - tag.header.timestamp + 10
+    delta = ((lastTags.last() as FlvTag).header.timestamp - tag.header.timestamp + 10).toInt()
   }
 
   fun correctTs(tag: FlvTag): FlvTag {
-    if (delta == 0L) {
+    if (delta == 0) {
       return tag
     }
     return tag.copy(header = tag.header.copy(timestamp = tag.header.timestamp + delta))
